Update contract deployment and async sleep

+++ light-clients/icsxx-ethereum/src/abi.rs
@@ -127,7 +127,7 @@ impl<H> From<ClientState<H>> for EthereumClientPrimitivesClientState {
 				.frozen_height
 				.map(Into::into)
 				.unwrap_or_else(|| HeightData { revision_number: 0, revision_height: 0 }),
-			latestHeight: value.latest_height,
+			latestHeight: value.latest_height as u32, // TODO: make latest_height u64 everywhere
 		}
 	}
 }
@@ -143,7 +143,7 @@ impl<H> TryFrom<EthereumClientPrimitivesClientState> for ClientState<H> {
 			} else {
 				Some(value.frozenHeight.into())
 			},
-			latest_height: value.latestHeight,
+			latest_height: value.latestHeight as u64,
 			_phantom: Default::default(),
 		})
 	}
